Digital_Assistant/requirements.txt
absl-py==0.10.0
aenum==3.1.0
aio-pika==6.8.0
aiocontextvars==0.2.2
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
appdirs==1.4.4
APScheduler==3.6.3
arrow==0.17.0
asgiref==3.3.4
astropy==4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
backoff==1.10.0
bidict==0.21.2
blis==0.4.1
boto3==1.17.43
botocore==1.20.43
cachelib==0.6.0
cachetools==4.2.1
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cloudpickle==1.4.1
colorama==0.4.4
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
configparser==4.0.2
contextvars==2.4
cryptography==3.3.2
cycler==0.10.0
cymem==2.0.5
dateparser==1.0.0
Deprecated==1.2.13
distlib==0.3.2
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
et-xmlfile==1.1.0
fbmessenger==6.0.0
filelock==3.0.12
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
ftfy==5.9
future==0.18.2
gast==0.3.3
gevent==21.1.2
google-api-core==2.0.1
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-cloud-speech==2.8.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.39.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.19.1
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.1
idna==2.10
idna-ssl==1.1.0
immutables==0.15
importlib-metadata==3.10.0
importlib-resources==5.1.4
investpy==1.0.7
ipdb==0.13.8
ipython-genutils==0.2.0
itsdangerous==2.0.1
jaeger-client==4.6.1
jedi==0.18.0
jellyfish==0.8.9
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.2
jsonschema==3.2.0
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libcst==0.3.20
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mattermostwrapper==2.2
multidict==4.7.6
murmurhash==1.0.5
mypy-extensions==0.4.3
networkx==2.5
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
oci==2.38.1
onetimepass==1.0.1
openpyxl==3.0.9
opentelemetry-api==1.5.0
opentelemetry-exporter-otlp==1.5.0
opentelemetry-exporter-otlp-proto-grpc==1.5.0
opentelemetry-instrumentation==0.24b0
opentelemetry-proto==1.5.0
opentelemetry-sdk==1.5.0
opentelemetry-semantic-conventions==0.24b0
opentracing==2.4.0
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.1.5
parsedatetime==2.6
parso==0.8.2
pathy==0.4.0
pickleshare==0.7.5
Pillow==8.2.0
plac==1.1.3
preshed==3.0.5
progressbar2==3.53.3
prompt-toolkit==2.0.10
proto-plus==1.19.0
protobuf==3.15.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pydot==1.4.2
Pygments==2.9.0
pykwalify==1.7.0
pymongo==3.10.1
pyOpenSSL==19.1.0
pyotp==2.6.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
pyserial==3.5
pyTelegramBotAPI==3.7.7
python-can==3.3.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.1
python-utils==2.5.6
pytz==2020.5
PyYAML==5.3.1
questionary==1.5.2
rasa==2.4.3
rasa-sdk==2.8.4
redis==3.5.3
regex==2020.9.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
rfc3986==1.4.0
rocketchat-API==1.9.1
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
sacremoses==0.0.44
sanic==20.9.0
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.95
sentry-sdk==0.19.5
six==1.14.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smart-open==3.0.0
sniffio==1.2.0
spacy==2.2.4
spacy-alignments==0.7.2
spacy-legacy==3.0.2
SpeechRecognition==3.8.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
srsly==1.0.5
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.4.0
threadloop==1.0.2
threadpoolctl==2.1.0
thrift==0.13.0
tokenizers==0.7.0
toml==0.10.2
torch==1.8.1
torchcontrib==0.0.2
tornado==6.1
totp==1.3.0
tqdm==4.50.2
traitlets==4.3.3
transformers==2.11.0
twilio==6.45.4
typeguard==2.12.0
typer==0.3.2
typing-extensions==3.7.4.2
typing-inspect==0.7.1
tzlocal==2.1
ujson==3.2.0
Unidecode==1.3.2
urllib3==1.26.5
urlvalidator==0.1
virtualenv==20.4.7
wasabi==0.8.2
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==2.0.1
wrapt==1.12.1
xlrd==2.0.1
yarl==1.5.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
